Move dapol node code to node_types
